Updated Betting Button Pressability
Updated the Betting Button Pressability so that deal and change bet buttons start disabled, deal can only be enabled if bet > 0, cash out sets bet to 0, and bet becomes 0 after a round if the amount of the bet is above the amount of remaining funds.  created a Player.refreshBetDisplay method for determining which betting related buttons can be active when betting is allowed.